Fix bug preventing custom maps in netgames
This was happening because the title screen code was resetting the
`currgame` variable in order to draw the original ground terrain.
Instead, change the `swground` function so that it can draw arbitrary
grounds without needing to change the variable.

A minor tweak here as well to change `asynhost` so that it is no longer
a static buffer.