Fix bug where targets can blow themselves up
This is rather hard to trigger, but it is sometimes possible for hangars
to blow themselves up by shooting their own flag. This happens because
bullets are spawned inside the target and we rely on their initial
momentum to push them outside it before the first collision check
happens. However, if the momentum isn't quite enough then it might not
escape. I've verified this bug also exists in the Network/Author's
edition.

In fixing this I'm trying to avoid significant or noticeable changes the
original behavior (no change to momentum for example). This approach
retains the existing behavior for the majority of bullets that get
spawned; we only adjust the tiny fraction of bullets that cause the
collisions.

Part of this required some minor refactoring of the `colltest()`
function so that it can be used in the checks. The result is the
modified `CollisionTest()` function that just returns a boolean rather
than changing the "kill" arrays.