Fix bug with ground overlapping edges of sprites
This could be seen in several places where there were right-facing
cliffs; in the original level for example, on the slope at the left
side of the map.

This bug is found in the DOS Sopwith 2 (and all other versions that draw
a line for the ground), but not in Sopwith 1 where there is solid
ground. This is an artifact of the algorithm being used to draw the
ground; it was not actually an accurate depiction of the ground heights.
This fixes the overlap with a new algorithm that always draws the ground
at the correct height.