Allow targets larger than 16x16
The main things here are that more ground needs to be flattened if the
sprite is wider, and the Y position needs to be adjusted if the sprite
is taller; we calculate these based on the symbol dimensions.
Furthermore, when a target is destroyed we may need to adjust the
position if the new symbol is a different size, and the position that
bullets fire from needs to be calculated from the symbol size too.