Never draw ground targets "hovering" above ground
With taller ground targets the centerpoint of the object can be high
enough above the ground that there appears to be a gap between the
ground terrain and the pixel representing the object on the map. We need
to make sure that any ground objects always appear on the map to
actually be on the map.