Properly clip sprites against left-side screen edge
With this change, sprites no longer "pop in" suddenly when flying to the
left, but are correctly drawn partially on-screen when their X
coordinate is negative relative to the screen. Besides being a nice
minor visual enhancement, this may open the door to wider in-game
objects in the future - these were not very practical in the past
because the "pop in" would be more pronounced the wider the object.