Plumb in the new high score system
This provides a complete initial version of the high score system (#10).
The high score table is shown at the title screen after a few seconds,
and if the player attains a new high score, they are prompted to enter
their name and it is added to the table. However, there is not yet any
persistence of the high score table: the new high scores are lost when
the game exits.