Fix failing `make distcheck` install step
The `make distcheck` command goes through the full series of steps of
extracting, building and installing the distribution package, but the
final step will fail if we try to use the default install path for the
high scores file (in `/var/games`). Instead, redirect it to install into
the same temporary directory that it uses for installing the other
files.