Fix "tailspin" nosedive animation
This fixes a minor annoyance that I noticed a while back: when the plane
is crashing and heading directly downward towards the ground, the game
animates the plane to show it in a tailspin. However, the animation for
the tailspin was only two frames, so it would actually appear to just be
jerking back and forward rather than spinning.

This adds an extra two frames to give the proper tailspin effect. An
alternative I considered here was to just mirror the existing two
frames, but I decided on this approach instead, because (1) the first
and third frames are not actually mirrors (they depict the bottom and
top of the plane); and (2) this approach potentially gives more
flexibility for modders.

In terms of the project philosophy, this is a change to the game, but I
judge it to be a very minor one with no significant impact on the
gameplay.