Don't install game as setgid to the `games` group
This fixes #57: we were installing the game as setgid to the `games`
group to control access to the system-wide high scores file (the same
mechanism used by other games such as Nethack). However, when running
under a Wayland desktop, SDL apparently uses Gtk+ and the Gtk+ team have
this ideological opposition to anything being setgid, to the extent that
Gtk+ will abort the program on startup if it detects that it is
installed as setgid.

This changes the installation hooks so that the game is no longer
installed as `setgid`, and `hiscores.txt` is now installed as
world-writable. This does mean that anyone can now cheat by just editing
it to put whatever name and score they want in the high scores table,
which is a shame, but ultimately harmless.