Fix negative score killing animals in multiplayer
As with prior commit commit 86d93298c858fa5, this fixes scoring in
multiplayer mode to be reversed when killing animals, matching the
single-player behavior.

However, currently it is always player 1 that gets punished when an
animal is killed, even if it was player 2 that did it. Scoring in
multiplayer mode needs to be better thought through in general, but for
the time I've filed a bug (#58) and added a TODO note.